feat: add tool CLI and plugin registry docs
