+++ cmd/agentry/plugin.go
@@ -36,24 +36,3 @@ func runPluginCmd(args []string) {
 		os.Exit(1)
 	}
 }
-
-func runToolCmd(args []string) {
-	if len(args) == 0 {
-		fmt.Println("Usage: agentry tool <command> [args]")
-		os.Exit(1)
-	}
-	switch args[0] {
-	case "init":
-		if len(args) < 2 {
-			fmt.Println("Usage: agentry tool init <name>")
-			os.Exit(1)
-		}
-		if err := plugin.InitTool(args[1]); err != nil {
-			fmt.Printf("init failed: %v\n", err)
-			os.Exit(1)
-		}
-	default:
-		fmt.Printf("unknown tool command %s\n", args[0])
-		os.Exit(1)
-	}
-}
